fix: invoiceDate format and unitPrice decimal places

+++ src/main/kotlin/com/mingshu/ecommerce/dto/SearchResponse.kt
@@ -2,9 +2,6 @@ package com.mingshu.ecommerce.dto
 
 
 class SearchResponse {
-
-	var invoices: List<InvoiceDto> = listOf()
-
-	var totalElements: Int =0
-
+    var invoices: List<InvoiceDto> = listOf()
+    var totalElements: Long = 0
 }
